Melhoria na arquitetura do projeto
