+++ backend/repos/mongo.py
@@ -12,6 +12,10 @@ class MongoRepo:
         row_id = self.collection.insert_one(row).inserted_id
         return row_id
 
+    def select(self, query_dict):
+        row_list = self.collection.find(query_dict)
+        return row_list
+
     def select_one(self, query_dict):
         row_dict = self.collection.find_one(query_dict)
         return row_dict
